cron job to refresh reading room cache
