+++ Dockerfile.prod
@@ -7,6 +7,7 @@ RUN apt-get update
 RUN apt-get install --yes apache2 apache2-dev 
 RUN apt-get install --yes postgresql
 RUN apt-get -y install python3-pip
+RUN apt-get -y install cron
 RUN pip install --upgrade pip
 
 RUN wget https://github.com/GrahamDumpleton/mod_wsgi/archive/refs/tags/4.9.0.tar.gz \
@@ -31,7 +32,9 @@ WORKDIR /var/www/request-broker
 RUN pip install -r requirements.txt
 RUN ./manage.py collectstatic
 
-EXPOSE ${APPLICATION_PORT}
-ADD . /code/
+COPY --chmod=644 crontab /etc/cron.d/crontab
+RUN crontab /etc/cron.d/crontab
 
-ENTRYPOINT ["/code/entrypoint.prod.sh"]
+EXPOSE ${APPLICATION_PORT}
+
+ENTRYPOINT ["./entrypoint.prod.sh"]
