+++ crontab
@@ -1,0 +1,2 @@
+*/5 * * * * python /var/www/request-broker/manage.py runcrons >/proc/1/fd/1 2>/proc/1/fd/2
+
