+++ requirements.in
@@ -2,6 +2,7 @@ ArchivesSnake~=0.9
 Django~=4.1
 django-cors-headers~=3.13
 django-csp~=3.7
+django4-cron~=0.5
 djangorestframework~=3.13
 inflect~=6.0.0
 ordered-set~=4.1
